fix: Move commands.Bot listener methods to discord.Bot This allows listeners to work in discord.Cog
